:recycle: Separate common methods to DF_helper.rb
