+++ lib/red_amber.rb
@@ -4,6 +4,7 @@ require 'arrow'
 require 'rover-df'
 
 require_relative 'red_amber/data_frame_displayable'
+require_relative 'red_amber/data_frame_helper'
 require_relative 'red_amber/data_frame_selectable'
 require_relative 'red_amber/data_frame'
 require_relative 'red_amber/vector_functions'
